city and hotel with relation with distance

+++ src/main/java/com/example/cityHotel/model/locations.java
@@ -1,26 +1,26 @@
 package com.example.cityHotel.model;
 
 public class locations {
-    private long latitude;
-    private long longitude;
+    private double latitude;
+    private double longitude;
 
     public locations() {
     }
 
-    public locations(long latitude, long longitude) {
+    public locations(double latitude, double longitude) {
         this.latitude = latitude;
         this.longitude = longitude;
     }
 
-    public long getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(long latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public long getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
